requirements.txt
discord.py==1.6.0
asyncpg==0.22.0
python-dotenv==0.15.0
fuzzywuzzy==0.18.0